% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Population of Countries from 1950 to 2020}
\usage{
CountryPopulation(data, CountryName)
}
\arguments{
\item{The}{name of a country as a string as in `Italy`}
}
\value{
A graph of the country's population from 1950 to 2020
}
\description{
This function will create a graph of a country's population over the
time period 1950 to 2020.Populations are estimates and are in thousands.
}
\details{
If the name of the country is not present in the data set the function
will return "No data available".
}
\examples{
CountryPopulation(WorldPopulation, "Burundi")
CountryPopulation(WorldPopulation, "Kuwait")
}
