% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{This is a data set containing information on attendance at all
World Cup events from 1930 to 1922}
\format{
a data frame with 22 observations and 4 columns
\describe{
  \item{WorldCup}{The host country and year}
  \item{Matches}{The number of matches over the course of the event}
  \item{Totalattendance}{The total attendance at all matches combined}
  \item{Averageattendance}{The average attendance per match as the total
  divided by the number of matches}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
This is a data set containing information on attendance at all
World Cup events from 1930 to 1922
}
\keyword{datasets}
