% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population by Country for the Years 1950 to 2020}
\format{
A data frame with 235 observations and 72 columns.
\describe{
   \item{Country Name}{This is the name of the country using UN designations}
   \item{Years}{Each subsequent column lists the estimated population
   of the country in thousands each year from 1950 to 2020}
}
}
\usage{
WorldPopulation
}
\description{
This data set contains population estimates from the United Nations
Population Division.
Population numbers are combined for both sexes, and represent
population in thousands.
Each country is listed by its official UN designation.
}
\keyword{datasets}
